Organized folders and codes
I have just re-organized the folders, and separated it as the the first part and the second part. The first part contains two codes, one was just the starting code, the other one is complete code for the basic functionality of the part 1(so bonus is remaining). And the part II is empty, which we can start working on simultaneously with the bonus part.